+++ src/spooq2/transformer/enum_cleaner.py
@@ -2,15 +2,17 @@ import sys
 import pyspark.sql.functions as F, types as T
 from pyspark.sql.column import Column
 
-from .transformer import Transformer
+from .base_cleaner import BaseCleaner
 
 
-class EnumCleaner(Transformer):
+class EnumCleaner(BaseCleaner):
     """
     Cleanses a dataframe based on lists of allowed|disallowed values.
 
-    Example
-    -------
+    Examples
+    --------
+    >>> from spooq2.transformer import EnumCleaner
+    >>>
     >>> transformer = EnumCleaner(
     >>>     cleaning_definitions={
     >>>         "status": {
@@ -19,16 +21,52 @@ class EnumCleaner(Transformer):
     >>>         "version": {
     >>>             "elements": ["", "None", "none", "null", "NULL"],
     >>>             "mode": "disallow",
-    >>>             "default": None
+    >>>             "default": None,
     >>>         },
     >>>     }
     >>> )
+
+    >>> from spooq2.transformer import EnumCleaner
+    >>> from pyspark.sql import Row
+    >>>
+    >>> input_df = spark.createDataFrame([
+    >>>     Row(a="stay", b="positive"),
+    >>>     Row(a="stay", b="negative"),
+    >>>     Row(a="stay", b="positive"),
+    >>> ])
+    >>> transformer = EnumCleaner(
+    >>>     cleaning_definitions={
+    >>>         "b": {
+    >>>             "elements": ["positive"],
+    >>>             "mode": "allow",
+    >>>         }
+    >>>     },
+    >>>     column_to_log_cleansed_values="cleansed_values_enum"
+    >>> )
+    >>> output_df = transformer.transform(input_df)
+    >>> output_df.show()
+    +----+--------+--------------------+
+    |   a|       b|cleansed_values_enum|
+    +----+--------+--------------------+
+    |stay|positive|                  []|
+    |stay|    null|          [negative]|
+    |stay|positive|                  []|
+    +----+--------+--------------------+
+    >>> output_df.printSchema()
+    root
+     |-- a: string (nullable = true)
+     |-- b: string (nullable = true)
+     |-- cleansed_values_enum: struct (nullable = false)
+     |    |-- b: string (nullable = true)
 
     Parameters
     ----------
     cleaning_definitions : :py:class:`dict`
         Dictionary containing column names and respective cleansing rules
 
+    column_to_log_cleansed_values : :any:`str`, Defaults to None
+        Defines a column in which the original (uncleansed) value will be stored in case of cleansing. If no column
+        name is given, nothing will be logged.
 
     Note
     ----
@@ -63,13 +101,17 @@ class EnumCleaner(Transformer):
     If you want to replace Null values you should use the method ~pyspark.sql.DataFrame.fillna from Spark.
     """
 
-    def __init__(self, cleaning_definitions={}):
-        super().__init__()
-        self.cleaning_definitions = cleaning_definitions
+    def __init__(self, cleaning_definitions={}, column_to_log_cleansed_values=None):
+        super().__init__(cleaning_definitions, column_to_log_cleansed_values)
         self.logger.debug("Enumeration List: " + str(self.cleaning_definitions))
+        self.TEMPORARY_COLUMNS_PREFIX = "9b7798529fef529c8f2586be7ca43a66"  # SHA1 hash of "EnumCleaner"
 
     def transform(self, input_df):
         self.logger.debug("input_df Schema: " + input_df._jdf.schema().treeString())
+        if self.column_to_log_cleansed_values:
+            column_names_to_clean = self.cleaning_definitions.keys()
+            temporary_column_names = self._get_temporary_column_names(column_names_to_clean)
+            input_df = self._add_temporary_columns(input_df, column_names_to_clean, temporary_column_names)
 
         for column_name, cleaning_definition in list(self.cleaning_definitions.items()):
             self.logger.debug(f"Cleaning Definition for Column {column_name}: {str(cleaning_definition)}")
@@ -109,5 +151,8 @@ class EnumCleaner(Transformer):
                 raise ValueError(
                     f"Only the following modes are supported by EnumCleaner: 'allow' and 'disallow'."
                 )
+
+        if self.column_to_log_cleansed_values:
+            input_df = self._log_cleansed_values(input_df, column_names_to_clean, temporary_column_names)
 
         return input_df