+++ tests/unit/transformer/test_exploder.py
@@ -30,13 +30,11 @@ class TestExploding(object):
     def default_params(self):
         return {"path_to_array": "attributes.friends", "exploded_elem_name": "friend"}
 
-    @pytest.mark.slow
     def test_count(self, input_df, default_params):
         expected_count = input_df.select(sql_funcs.explode(input_df[default_params["path_to_array"]])).count()
         actual_count = Exploder(**default_params).transform(input_df).count()
         assert expected_count == actual_count
 
-    @pytest.mark.slow
     def test_exploded_array_is_added(self, input_df, default_params):
         transformer = Exploder(**default_params)
         expected_columns = set(input_df.columns + [default_params["exploded_elem_name"]])
@@ -44,7 +42,6 @@ class TestExploding(object):
 
         assert expected_columns == actual_columns
 
-    @pytest.mark.slow
     def test_array_is_converted_to_struct(self, input_df, default_params):
         def get_data_type_of_column(df, path=["attributes"]):
             record = df.first().asDict(recursive=True)