+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog
 =========
 
+3.2.0 (2021-04-13)
+-------------------
+* [MOD] add functionality to log cleansed values into separate struct column (column_to_log_cleansed_values)
+* [MOD] add ignore_ambiguous_columns to Mapper
+* [MOD] log spooq version when importing
+* [REM] Drop separate spark package (bin-folder) as pip package can now handle all tests as well
+* [ADD] Github action to test on label (test-it) or merge into master
+
 3.1.0 (2021-01-27)
 -------------------
 * [ADD] EnumCleaner Transformer